Update transactions count calculation

+++ calculations/src/main/java/com/mapofzones/calculations/transactions/repository/mongo/domain/TxsChart.java
@@ -5,10 +5,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Getter
 @Setter
@@ -31,7 +29,7 @@ public class TxsChart {
             this.zone = zone;
         }
         private String zone;
-        private BigInteger totalTxsCount;
+        private Integer totalTxsCount;
         private List<ChartItem> chart;
 
         @Getter
